Melhoria na legibilidade dos outputs em txt. Configurando queries para ignorar IPs indesejados. Alteração no nome dos arquivos de output TXT para algo mais coeso com as suas funcionalides
